+++ ts/packages/knowPro/src/index.ts
@@ -5,4 +5,6 @@ export * from "./import.js";
 export * from "./dataFormat.js";
 export * from "./conversation.js";
 export * from "./conversationIndex.js";
+export * from "./secondaryIndexes.js";
+export * from "./relatedTermsIndex.js";
 export * from "./search.js";
